functions/requirements.txt
flask
firebase-admin
numpy
pandas
firebase-functions
twilio
anthropic
reportlab
python-dotenv
